requirement.txt
selenium
webdriver-manager
requests
python-dotenv
flask